backend/requirements.txt
black==23.3.0
Django==3.2.3
django-filter==23.2
djangorestframework==3.12.4
djoser==2.2.0
flake8==6.0.0
flake8-isort==6.0.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
requests==2.26.0
webcolors==1.11.1
